+++ bigbuy/api.py
@@ -5,7 +5,6 @@ Official documentation for Bigbuy API endpoints can be found at: https://api.big
 """
 import base64
 import mimetypes
-import warnings
 from http.cookiejar import DefaultCookiePolicy
 from typing import Optional, Dict, Union, Iterable, List, cast, Any
 
@@ -194,14 +193,6 @@ class BigBuy(APISession):
         """Returns all products' information."""
         return self.get_json_api('catalog/productsinformation', params=params)
 
-    def get_products_stock(self, **params):
-        """Returns all products stock."""
-        warnings.warn("The endpoint used by `get_products_stock` has been deprecated by BigBuy"
-                      " and will be removed on April 1st, 2024."
-                      " Use `get_products_stock_by_handling_days` instead.",
-                      DeprecationWarning)
-        return self.get_json_api('catalog/productsstock', params=params)
-
     def get_products_stock_by_handling_days(self, **params):
         """Returns all products stock by handling days."""
         return self.get_json_api('catalog/productsstockbyhandlingdays', params=params)
@@ -219,14 +210,6 @@ class BigBuy(APISession):
         """Lists all product tags."""
         return self.get_json_api('catalog/productstags', params=params)
 
-    def get_product_stock(self, product_id: Id, **params):
-        """Get a single product stock."""
-        warnings.warn("The endpoint used by `get_product_stock` has been deprecated by BigBuy"
-                      " and will be removed on April 1st, 2024."
-                      " Use `get_product_stock_by_handling_days` instead.",
-                      DeprecationWarning)
-        return self.get_json_api(f'catalog/productstock/{product_id}', params=params)
-
     def get_product_stock_by_handling_days(self, product_id: Id, **params):
         """Get a single product stock by handling days."""
         return self.get_json_api(f'catalog/productstockbyhandlingdays/{product_id}', params=params)
@@ -234,14 +217,6 @@ class BigBuy(APISession):
     def get_products_variations(self, **params):
         """Returns all products variations."""
         return self.get_json_api('catalog/productsvariations', params=params)
-
-    def get_products_variations_stock(self, **params):
-        """Returns all products variations stock."""
-        warnings.warn("The endpoint used by `get_products_variations_stock` has been deprecated by BigBuy"
-                      " and will be removed on April 1st, 2024."
-                      " Use `get_products_variations_stock_by_handling_days` instead.",
-                      DeprecationWarning)
-        return self.get_json_api('catalog/productsvariationsstock', params=params)
 
     def get_products_variations_stock_by_handling_days(self, **params):
         """Returns all products variations stock by handling days."""
@@ -254,14 +229,6 @@ class BigBuy(APISession):
     def get_product_variations(self, product_id: Id, **params):
         """Get a single Product variations."""
         return self.get_json_api(f'catalog/productvariations/{product_id}', params=params)
-
-    def get_product_variations_stock(self, product_id: Id, **params):
-        """Get a single product variation stock."""
-        warnings.warn("The endpoint used by `get_product_variations_stock` has been deprecated by BigBuy"
-                      " and will be removed on April 1st, 2024."
-                      " Use `get_product_variations_stock_by_handling_days` instead.",
-                      DeprecationWarning)
-        return self.get_json_api(f'catalog/productvariationsstock/{product_id}', params=params)
 
     def get_product_variations_stock_by_handling_days(self, product_id: Id, **params):
         """Get a single product variation's stocks by handling days."""